pyproject.toml
[project]
name = "utano-checker"
version = "0.1.0"
description = "YouTube setlist analyzer"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.1", # .env用
    "google-api-python-client>=2.189.0", # Google API用
]
